+++ app/controllers/decidim_controller.rb
@@ -4,5 +4,4 @@
 # entry point, but you can change what controller it inherits from
 # so you can customize some methods.
 class DecidimController < ApplicationController
-  include Decidim::NeedsUserExtension
 end
